# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LearnVault API
  version: 1.0.0
  description: Backend API for LearnVault frontend and integrations.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ZodIssue:
      type: object
      required:
        - code
        - path
        - message
      properties:
        code:
          type: string
          example: invalid_type
        path:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          example:
            - walletAddress
        message:
          type: string
          example: Required
    ValidationErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ZodIssue"
...
  responses:
    BadRequestError:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/proposals/{proposalId}/comments:
    get:
      summary: Fetch comments for a proposal
      tags:
        - Comments
      parameters:
        - in: path
          name: proposalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
  /api/comments:
    post:
      summary: Post a new comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Comment created
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          description: Rate limited
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/comments/{id}:
    patch:
      summary: Edit own comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Comment updated
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete own comment (soft delete)
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/comments/{id}/vote:
    put:
      summary: Upvote or downvote a comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum:
                    - upvote
                    - downvote
      responses:
        "200":
          description: Vote recorded
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/comments/{id}/pin:
    put:
      summary: Pin a comment (proposal author only)
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pinned
              properties:
                pinned:
                  type: boolean
      responses:
        "200":
          description: Pin updated
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...